[BEAM-1871] Create new GCP core module package and move several GCP related classes from beam-sdks-java-core over.
